Table clean up so far
